requirements.txt
yt-dlp
openai-whisper
torch
google-generativeai
python-dotenv
fastapi
uvicorn